Changed collection update syntax to use verbs.
