+++ RavenMigrations/RavenMigrations.csproj
@@ -59,6 +59,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RavenMigrationHelpers.cs" />
     <Compile Include="Runner.cs" />
+    <Compile Include="Verbs\Alter.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
